Add support for SSH agent forwarding.
This allows use of private dependencies in builds by forwarding the SSH
agent to the docker container. This can be configured either via
`build.env`, `target.(...).env`, or `ENV_SSH_AGENT_PASSTHROUGH`. See the
crate https://github.com/Alexhuszagh/ssh_dependency for an example how
to configure this.

It works by mounting `$SSH_AUTH_SOCK` to `/ssh-agent` and then providing
an environment variable `SSH_AUTH_SOCK=/ssh-agent` to the container.
Note that this is not tested on Windows or macOS, only Linux right now,
so this might not work robustly.