Added JSON-based changes to auto-generate the changelog.
Solves merge conflicts between changelogs, by using sorting and
type-based merging for the changelog from a series of JSON files. This
allows conflict-free generation of a changelog using JSON-based changes,
stored in `.changes`.

Each entry can be object or array-based: the array must contain objects,
where each object is a different change. The entries are added to the
changelog, in sorted order, and the validity of the changelog is
verified.

To ensure fidelity of the changelog changes, the `no changelog` flag or
at least 1 change (either removed, or added/dified and almovidated) for
a JSON file under `.changes` must occur.

A `pre-release-hook` has been added so all these changes are automerged
into the CHANGELOG prior to release.